% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{bin_by_order}
\alias{bin_by_order}
\title{show data distribution among orders}
\usage{
bin_by_order(data)
}
\arguments{
\item{data}{data.frame. must contain a column named \code{order}, such as \code{omics}.}
}
\value{
\code{NULL}
}
\description{
show data distribution among orders
}
