% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{bind_taxonomy}
\alias{bind_taxonomy}
\title{bind columns of taxonomy information}
\usage{
bind_taxonomy(data)
}
\arguments{
\item{data}{data.frame. must contain a column named \code{species}, whose first
word is used to match genus name (in case of `'genus1 Sp'``). see
\link{extract_genus}}

\item{taxonomy}{data.frame. taxonomy database to be used.}
}
\value{
data.frame
}
\description{
bind columns of taxonomy information
}
\examples{
aves::bind_taxonomy(aves::genome)

}
