% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{diff_order}
\alias{diff_order}
\title{preserve only the first row for same species}
\usage{
diff_order(data, Source)
}
\arguments{
\item{data}{data.frame. must contain a column named \code{order}, such as \code{omics}.}

\item{Source}{string. such as \code{'SRA'} or \code{'genome'}}

\item{data}{data.frame. must contain a column named \code{order}, such as \code{omics}.}
}
\value{
data.frame.

\code{NULL}
}
\description{
preserve only the first row for same species

show orders significantly influenced (e.g., 2 data becomes 3 data) by a certain source
}
\examples{
aves::uniq_species(aves::rna)
}
