% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{extract_genus}
\alias{extract_genus}
\title{extract the first word (only consists of letters)}
\usage{
extract_genus(name)
}
\arguments{
\item{name}{character. can be scientific name or \code{name} of aves's data.frame}
}
\value{
character
}
\description{
used to extract genus
}
\examples{
    aves::extract_genus('Gallus gallus')
    aves::extract_genus('Gallus_gallus_2_TSA')


}
