% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hamstr_command}
\alias{hamstr_command}
\title{generate hamstr command}
\usage{
hamstr_command(taxon, hmmpath = "/path/to/data/aves/hamstr", hmmer_top = 10L,
  dir = "hamstr", hmm_set = "aves", threads = getOption("mc.cores"))
}
\arguments{
\item{taxon}{character. taxon name, see \link{extract_genus}}

\item{hmmer_top}{integer scalar. passed on to \code{hit_limit} option. specify
\code{NULL} to disable that option.}

\item{dir}{string. folder where the input file lies and output file will be
put}

\item{hmm_set}{character.  passed on to \code{hmmset} option.}

\item{threads}{integer scalar. how many tasks to run in parallel}

\item{dna}{logical scalar. Whether the input sequence is EST or protein}
}
\value{
character
}
\description{
generate hamstr command
}
\section{to do}{
 remove \code{mod} \code{mod.tc} if older
}

\examples{
{
    aves::hamstr_command('Arborophila_rufipectus_SRA_1')
    aves::hamstr_command('Gallus_Gallus', F, 'hamstr-ext', NULL, 'Gallus_Gallus')
}
}
