% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{rectify_ncbi_species}
\alias{rectify_ncbi_species}
\title{correct species name error in NCBI genome & assembly database}
\usage{
rectify_ncbi_species(species)
}
\arguments{
\item{species}{character. species scientific name}
}
\value{
species
}
\description{
correct species name error in NCBI genome & assembly database
}
