% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{summary_as_yaml}
\alias{summary_as_yaml}
\title{summary data distribution using the YAML format}
\usage{
summary_as_yaml(data)
}
\arguments{
\item{data}{data.frame. must contain a column named \code{species}, such as
\code{aves::protein}. would be transformed by \link{bind_taxonomy} if doesn't contain a
column named \code{order}}
}
\value{
string. content of the yaml file. I suggest \code{\link[readr:write_file]{readr::write_file()}}
}
\description{
summary data distribution using the YAML format
}
\examples{
aves::summary_as_yaml(aves::protein)
}
