% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trimmomatic_command}
\alias{trimmomatic_command}
\title{command for running trimmomatic}
\usage{
trimmomatic_command(id, adapter = "TruSeq3", extra = "")
}
\arguments{
\item{id}{string. SRA accession}

\item{adapter}{string. \code{'TruSeq3'} or \code{'TruSeq2'}}

\item{extra}{}
}
\value{
string.
}
\description{
command for running trimmomatic
}
\section{Adapter}{
 according to the "Overrepresented Sequences" report of FASTQC
\enumerate{
\item "Illumina Single End" or "Illumina Paired End" : TruSeq2
\item "TruSeq Universal Adapter" or "TruSeq Adapter, Index ..." : TruSeq3
}
}

\examples{

}
